Victor added posibility for Company to add service,invoice. Display all his services

+++ src/main/java/com/endava/service_system/model/InvoiceFilter.java
@@ -19,10 +19,10 @@ public class InvoiceFilter {
     private String currentUserUsername;
     //complete only if userType == User
     private String companyTitle;
-    private Integer companyId;
+    private Long companyId;
     private InvoiceStatus invoiceStatus;
     private String categoryName;
-    private Integer categoryId;
+    private Long categoryId;
     private Integer size;
     private Integer page;
     private Sort.Direction orderByDueDateDirection;
